UIWRKFLOW-7: Extract parts of the upload file into a useUploadMultipart() hook.
