+++ src/hooks/index.ts
@@ -6,3 +6,4 @@ export * from './useLocalStorageToggle';
 export * from './useLists';
 export * from './usePrevious';
 export * from './useSearch';
+export * from './useUpload';
