+++ src/hooks/useUpload/index.ts
@@ -1,0 +1,1 @@
+export { useUploadMultipart } from './useUploadMultipart';
